fix: prevent leaking AWS credentials in workflow
Login into AWS after the build is completed, so the authorization
environment variables aren't present. Do so thru the GitHub OIDC
provider for logging into AWS to minimize the lifetime of the
credentials, minimizing the impact of a leak, should it happen. Unpin
the AWS action version and bump it to latest v4.